Account feature
Add User and authentification service.
User file isolation
Frontend User Authentification and manage page
M2M token security.
Unit Test

+++ frontendApp/src/app/models/data-mapper-types.ts
@@ -227,7 +227,7 @@ export class DataMapperUtils {
     if (/(_id|Id|id)$/i.test(fieldName)) {
       return 'id';
     }
-    if (type === 'int64' || type === 'float64') {
+    if (type === 'int' || type === 'float') {
       if (fieldName === 'price' || fieldName === 'cost') {
         return 'currency';
       } else {
